+++ docs/changelog.rst
@@ -10,6 +10,9 @@ Changelog
 
 * Added an interactive client: `python -m websockets <uri>`
 
+* Changed the ``origins`` argument to represent the lack of an origin with
+  ``None`` rather than ``''``.
+
 * Improved error messages when a required HTTP header is missing.
 
 6.0